+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  - Added *FormValues.fieldname()* and *UiFormPluginValues.getFieldName()* methods.
  - Added *input-group* support for error states and display.
  - Fixed *input.error* and *input.error.visible* state and error clearing.
+ - Changed *UiFormComponent* config option *async* default value to *true*.
  - Updated documentation and dependencies.
 
 ## 0.6.0